LPS-161789 Moves class to taglib module
